Splits files up, new detail/ and reader/ namespaces
