+++ quasardb/qdb_client.cpp
@@ -55,5 +55,7 @@ PYBIND11_MODULE(quasardb, m)
     qdb::register_ts_reader(m);
 
     qdb::detail::register_ts_column(m);
+    qdb::reader::register_ts_value(m);
+    qdb::reader::register_ts_row(m);
 
 }
